refactor(bip324): consolidate parser files into mod.rs with separate tests file
